[core] Merge RenderStyle into Renderer::Impl
