+++ cmake/core-files.cmake
@@ -194,9 +194,6 @@ set(MBGL_CORE_FILES
     src/mbgl/renderer/render_source_observer.hpp
     src/mbgl/renderer/render_static_data.cpp
     src/mbgl/renderer/render_static_data.hpp
-    src/mbgl/renderer/render_style.cpp
-    src/mbgl/renderer/render_style.hpp
-    src/mbgl/renderer/render_style_observer.hpp
     src/mbgl/renderer/render_tile.cpp
     src/mbgl/renderer/render_tile.hpp
     src/mbgl/renderer/renderer.cpp
