--- src/mbgl/renderer/render_style_observer.hpp
@@ -1,14 +1,0 @@
-#pragma once
-
-#include <exception>
-
-namespace mbgl {
-
-class RenderStyleObserver {
-public:
-    virtual ~RenderStyleObserver() = default;
-    virtual void onInvalidate() {}
-    virtual void onResourceError(std::exception_ptr) {}
-};
-
-} // namespace mbgl
